+++ src/main/java/com/laptrinhjava/ShoppingCart/payload/response/ProductResponse.java
@@ -11,9 +11,9 @@ public class ProductResponse {
     private Long id;
     private String name;
     private String imageURL;
-    private Long price;
+    private Double price;
 
-    public ProductResponse(Long id, String name, String imageURL, Long price) {
+    public ProductResponse(Long id, String name, String imageURL, Double price) {
         this.id = id;
         this.name = name;
         this.imageURL = imageURL;
